+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.2.23
 
+- Fix `Glimmer::LibUI::timer {}` behavior so that the block return value will not affect repetition outcome if it is `Integer` (only Booleans affect it)
 - Fix Tetris down button for Windows (it was going too fast before making tetrominos invisible before they hit the bottom)
 - Fix Tetris double-downs (or multiple downs) happening after game over (it was firing an additional down timer after game over by mistake before)
 - Fix Tic-Tac-Toe text-size for Windows/Linux